#780 Refactor the cobol processor ot use the proper builder pattern.
